% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_features.R
\name{make_saf}
\alias{make_saf}
\title{SAF converter}
\usage{
make_saf(ann_file, exclude = FALSE)
}
\arguments{
\item{ann_file}{Either a GFF3 annotation file or a pre-built GFF cache (see \code{load_gff_cache}).}

\item{exclude}{A boolean to indicate whether or not to exclude rRNA/tRNA features}
}
\value{
A dataframe in Simplified Annotation Format
}
\description{
This function converts gff file into simplified annotation format with 5 columns, appropriate for use in featureCounts/tpm_norm_flagging.
}
\note{
By design this quantification filter excludes only tRNA and rRNA when
  \code{exclude = TRUE}, keeping all other annotated RNAs so they are counted.
  This is deliberately more permissive than \code{major_features()}
  (prediction), which excludes all pre-annotated ncRNAs to avoid
  re-discovering them.
}
