% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_file_editor.R
\name{peak_analysis}
\alias{peak_analysis}
\title{Peak checking for the second coverage threshold and width.}
\usage{
peak_analysis(View_line, high_cutoff, min_sRNA_length)
}
\arguments{
\item{View_line}{A line from a RleViews object.}

\item{high_cutoff}{An integer indicating the high coverage threshold value. Exclusive: coverage must be greater than this value.}

\item{min_sRNA_length}{An integer indicating the minimum sRNA length (peak width). Exclusive: the peak must carry more than this many positions above the high cut-off.}
}
\value{
Returns a RleViews line if it satisfies conditions.
}
\description{
This is a helper function that is used to examine whether the number of positions with coverage above the high cut-off value exceeds the given width. The positions need not be contiguous; contiguity is imposed by the low cut-off slice that produces the RleViews line.
}
